add a login page for ios
