+++ src/constants.js
@@ -1,5 +1,6 @@
 // eslint-disable-next-line import/prefer-default-export
 export const BASE_URL = 'http://localhost:3000/api/v1';
+export const IOS_PREFIX = 'unihack://';
 export const NOTIFICATION_TYPE_TECH_TALK = 'techTalk';
 export const NOTIFICATION_TYPE_SESSION = 'session';
 export const NOTIFICATION_TYPE_SPECIAL = 'mealsRafflesEtc';
